+++ crates/jmap-proto/src/references/resolve.rs
@@ -60,7 +60,7 @@ impl Response<'_> {
                 GetRequestMethod::Registry(request) => request.resolve_references(self)?,
             },
             RequestMethod::Set(request) => match request {
-                SetRequestMethod::Email(request) => request.resolve_references(self, 1, false)?,
+                SetRequestMethod::Email(request) => request.resolve_references(self, 2, false)?,
                 SetRequestMethod::Mailbox(request) => request.resolve_references(self, 1, false)?,
                 SetRequestMethod::Identity(request) => {
                     request.resolve_references(self, 1, false)?